% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_functions.R
\name{build_gbif_spatial_data}
\alias{build_gbif_spatial_data}
\title{Functions in this script:
\itemize{
\item build_gbif_spatial_data()
\item get_gbif_data()
\item get_gbif_occ_data()
\item build_gbif_spp()
\item wkt_string()
\item gbif_spatial()
Subset eligible species from GBIF data and reduce variables.
}}
\usage{
build_gbif_spatial_data(gbif_data, spp_list)
}
\arguments{
\item{gbif_data}{Spatial GBIF data from \code{\link[=get_gbif_data]{get_gbif_data()}}.}

\item{spp_list}{Species list that includes taxon ID from \code{\link[=get_taxonomies]{get_taxonomies()}}.
This is the list that is used to subset the spatial data.}
}
\value{
An \link[sf:sf]{sf::sf} object.
}
\description{
Functions in this script:
\itemize{
\item build_gbif_spatial_data()
\item get_gbif_data()
\item get_gbif_occ_data()
\item build_gbif_spp()
\item wkt_string()
\item gbif_spatial()
Subset eligible species from GBIF data and reduce variables.
}
}
\examples{
\dontrun{
library("psoSppEvals")

# Project directory path
t_path <- file.path("T:/path/to/project/directory/data")

# Pull data from existing GBIF query
gbif_dat <- get_gbif_data(gbif_key = '9999999-999999999999999', 
                          t_path = t_path)
# Summarize species
gbif_list <- build_gbif_spp(gbif_dat)
# Subset data
birds <- dplyr::filter(gbif_list, class == "Aves")
# Subset spatial data
gbif_birds <- build_gbif_spatial_data(gbif_dat, birds)
}
}
\seealso{
\code{\link[=get_gbif_data]{get_gbif_data()}}, \code{\link[=build_gbif_spp]{build_gbif_spp()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
