% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{build_ns_spp_list}
\alias{build_ns_spp_list}
\title{Build NatureServe Species List}
\usage{
build_ns_spp_list(ns_data)
}
\arguments{
\item{ns_data}{NatureServe data from \code{get_ns_state_list()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function builds a species list with global and state NatureServe
rankings using data returned from \code{get_ns_state_list()}.
}
\examples{
\dontrun{
library("psoSppEvals")
co_ns_data <- get_ns_state_list("CO")
co_ns_spp_list <- build_ns_spp_list(co_ns_data)
}
}
\seealso{
\code{\link[=get_ns_state_list]{get_ns_state_list()}}
}
