% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_quarto_params.R
\name{build_quarto_params}
\alias{build_quarto_params}
\title{Build Quarto Parameters for Automated Evaluation Templates}
\usage{
build_quarto_params(nat_known_list, output_path, target_unit_name)
}
\arguments{
\item{output_path}{The base directory where you want the reports saved.}

\item{target_unit_name}{Name of FS unit "e.g. Dixie National Forests".}

\item{spp_list}{Species list with taxon ID's and taxonomies from
\code{get_taxonomies()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function builds the parameters for automated reporting.
}
\examples{
\dontrun{
library('mpsgSE')
spp_list <- get_taxonomies(mpsgSE::sp_list_ex, 'scientific_name', 
                           correct = TRUE)
qmd_params <- build_quarto_params(spp_list, file.path('output', 'spp_evlas'), 
                                  "Smokey Bear National Forest")
}
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}
}
