% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{combine_ns_habs}
\alias{combine_ns_habs}
\title{Combine Two NatureServe Habitat Data Frames}
\usage{
combine_ns_habs(state_a_habitats, state_b_habitats)
}
\arguments{
\item{state_a_habitats}{Idaho NatureServe Habitats from this pipeline.}

\item{state_b_habitats}{Montana NatureServe Habitats from this pipeline}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function combines two habitat data frames returned from
\code{get_ns_habitat()}.
}
\examples{
\dontrun{
library("psoSppEvals")

# Colorado
ns_co <- get_ns_state_list("CO")
hab_co <- get_ns_habitat(ns_co)

# Wyoming
ns_wy <- get_ns_state_list("WY")
hab_wy <- get_ns_habitat(ns_wy)

# Combine CO and WY
habitats <- combine_ns_habs(hab_co, hab_wy)
}
}
\seealso{
\code{\link[=get_ns_habitat]{get_ns_habitat()}}
}
