% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{combine_ns_spp_lists}
\alias{combine_ns_spp_lists}
\title{Combine NatureServe Data from build_ns_spp_list()}
\usage{
combine_ns_spp_lists(ns_spp_1, ns_spp_2)
}
\arguments{
\item{ns_spp_1}{Data frame from \code{build_ns_spp_list()}.}

\item{ns_spp_2}{Data frame from \code{build_ns_spp_list()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function combines two species data frames returned from
\code{build_ns_spp_list()}.
}
\examples{
\dontrun{
library("mpsgSE")

# Colorado
ns_co <- get_ns_state_list("CO")
spp_co <- build_ns_spp_list(ns_co)

# Wyoming
ns_wy <- get_ns_state_list("WY")
spp_wy <- build_ns_spp_list(ns_wy)

# Combine CO and WY
spp_list <- combine_ns_spp_lists(spp_co, spp_wy)
}
}
\seealso{
\code{\link[=build_ns_spp_list]{build_ns_spp_list()}}
}
