% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{combine_ns_state_lists}
\alias{combine_ns_state_lists}
\title{Combine NatureServe Data from get_ns_state_list()}
\usage{
combine_ns_state_lists(ns_list_1, ns_list_2)
}
\arguments{
\item{ns_list_1}{State data from \code{get_ns_state_list()}.}

\item{ns_list_2}{State data from \code{get_ns_state_list()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function combines the state NatureServe lists returned from
\code{get_ns_state_list()} and returns a species list data frame.
}
\examples{
\dontrun{
library("psoSppEvals")

ns_co <- get_ns_state_list("CO")
ns_ks <- get_ns_state_list("KS")
ns_data <- combine_natureserve_data(ns_co, ns_ks)
}
}
\seealso{
\code{\link[=get_ns_state_list]{get_ns_state_list()}}
}
