% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfire_functions.R
\name{extract_landfire_evt}
\alias{extract_landfire_evt}
\title{Extract EVT Data to Points and Summarize by Taxa}
\usage{
extract_landfire_evt(spp_list, spp_pts_sf, lf_dir)
}
\arguments{
\item{spp_list}{List of species with taxon ID's from \code{get_taxonomies()}.}

\item{spp_pts_sf}{\link[sf:sf]{sf::sf} points object of species occurrences with taxon ID's
from \code{get_taxonomies()}.}

\item{lf_dir}{Path to LANDFIRE directory.}
}
\value{
A list.
}
\description{
This function extracts LANDFIRE EVT data to species occurrence data and then
summarizes the EVT data by taxa.
}
\examples{
library(mpsgSE)
}
