% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_functions.R
\name{get_gbif_occ_data}
\alias{get_gbif_occ_data}
\title{Get GBIF occurrence records for a list of species}
\usage{
get_gbif_occ_data(spp_list, spatial = TRUE, crs = "EPSG:4326")
}
\arguments{
\item{spp_list}{A list of species with taxon ID's from \code{get_taxonomies()}.}

\item{spatial}{Logical (TRUE/FALSE). Return spatial data. Default is TRUE.}

\item{crs}{Target coordinate reference system (CRS). Either and
\code{sf::st_crs()} object or accepted string (e.g. "EPSG:4326" or "NAD83").
Default is EPSG:4326 (WGS84).}
}
\value{
An \link[sf:sf]{sf::sf} object or a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function queries GBIF occurrence records for a species list using the
taxon ID from \code{get_taxonomies()} using \code{\link[rgbif:occ_search]{rgbif::occ_search()}}.
}
\examples{
\dontrun{
library(psoSppEvals)

# Species list with taxon ID's
spp <- get_taxonomies(psoSppEvals::sp_list_ex, correct = TRUE)

# Pull occurrence data
occ <- get_gbif_occ_data(spp)
}
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}, \code{\link[rgbif:occ_search]{rgbif::occ_search()}}
}
