% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_functions.R
\name{get_iucn_maps}
\alias{get_iucn_maps}
\title{Read IUCN Species Range Maps into R}
\usage{
get_iucn_maps(spp_list)
}
\arguments{
\item{spp_list}{A species list with taxon ID's from \code{get_taxonomies()}.}
}
\value{
An \link[sf:sf]{sf::sf} object.
}
\description{
This function reads spatial (polygon) range data from International Union for
Conservation of Nature (IUCN) into R. These shapefiles were downloaded
from IUCN (\url{https://www.iucnredlist.org/resources/spatial-data-download}),
and saved on the MPSG T-drive. This function will this function will fail
if you are not on a FS network.
}
\examples{
library('mpsgSE')

spp_list <- get_taxonomies(mpsgSE::sp_list_ex, correct = TRUE)

iucn_maps <- get_iucn_maps(spp_list)

}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}
}
