% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{get_ns_state_list}
\alias{get_ns_state_list}
\title{Import NatureServe Species Data}
\usage{
get_ns_state_list(state, taxonomy = TRUE, correct = FALSE)
}
\arguments{
\item{state}{A character with state code, example: "CO".}

\item{taxonomy}{Logical. TRUE if you want to return the data with
taxonomic information included.  This can take a long time for states
with large lists.}

\item{correct}{Logical. If TRUE, \code{correct_taxon_ids()} is used to correct
known issues with taxon ID's and scientific names. Default is FALSE.}
}
\value{
A \code{\link[=list]{list()}}
}
\description{
This function imports NatureServe species data for a given state into R.
}
\examples{
\dontrun{
library("mpsgSE")
co_ns_data_no_taxonid <- get_ns_state_list("CO", taxonomy = FALSE)
co_ns_data <- get_ns_state_list("CO")
}
}
