% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual_corrections}
\alias{manual_corrections}
\title{Scientific names with manually corrected taxon ID's}
\format{
A data frame of 2 observations and 3 variables
\describe{
\item{common_name}{Common species name.}
\item{scientific_name}{Scientific name of species.}
\item{taxon_id}{Taxon ID.}
}
}
\usage{
manual_corrections
}
\description{
Manually generated through trial and error
}
\keyword{datasets}
