% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_service_species_data_functions.R
\name{pull_edw_bio_data}
\alias{pull_edw_bio_data}
\title{Pull biological data from the Forest Service Eneterprise Data Warehouse}
\usage{
pull_edw_bio_data(aoa_sf, crs, write = FALSE, gdb_path, suffix = "3miBuffer")
}
\arguments{
\item{aoa_sf}{Simple feature (\code{sf}) object of the area of analysis.}

\item{crs}{Target coordinate reference system.}

\item{write}{Logical (TRUE/FALSE). Write the data to a geodatabase. Default
is FALSE.}

\item{gdb_path}{Optional. Character. Directory path to the geodatabase.}

\item{suffix}{Optional.}
}
\value{
A simple feature (\code{sf}) object.
}
\description{
This function is in active development. This function pulls species
observation data from the internal Forest Service ArcGIS REST Service
(\emph{arcn}) into R.
}
\note{
Aquatic biota observation data (\emph{EDW_Aquatic_Biota_Observations_01}) is not
currently reading into R.
}
\examples{
library("dplyr")
library("psoGIStools")
library("sf")
library("units")

crs <- "EPSG:26912"
admin_bndry <- read_edw_lyr("EDW_ForestSystemBoundaries_01", 1, "arcx", 
                            crs = crs) |> 
  dplyr::filter(forestname == "Dixie National Forest")
aoa <- sf::st_buffer(admin_bndry, dist = units::as_units(3, "mi"))
fs_edw <- pull_edw_bio_data(aoa, crs = crs)

}
