% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfire_functions.R
\name{pull_landfire_evt}
\alias{pull_landfire_evt}
\title{Functions in this script:
\itemize{
\item pull_landfire_evt()
\item extract_landfire_evt()
Download LANDFIRE Existing Vegetation Type (250EVT)
}}
\usage{
pull_landfire_evt(
  plan_area_sf,
  lf_dir,
  email_address,
  res = 30,
  product = "LF2024_EVT"
)
}
\arguments{
\item{plan_area_sf}{Plan Area \code{sf} polygon object.}

\item{lf_dir}{Directory path to save raster to.}

\item{email_address}{Email address. Passed on to \code{\link[rlandfire:landfireAPIv2]{rlandfire::landfireAPIv2()}}.}

\item{res}{Raster resolution. Default is 30.}

\item{product}{A character for the EVT product of interest. Default is
\emph{LF2024_EVT}. See https://lfps.usgs.gov/products for mor list of products.}
}
\value{
A list of two data frames summarizing area of EVT.
}
\description{
This function downloads and saves LANDFIRE Existing Vegetation Type (250EVT)
raster data and then summarizes the data by EVT Name and EVT Physiognomy
for the Plan Area (FS lands).
}
\examples{
\dontrun{
library(psoSppEvals)

message("Example coming soon. To a theater near you.")

}
}
