% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{read_edw_lyr}
\alias{read_edw_lyr}
\title{\strong{Deprecated}. Read spatial data from Forest Service ArcGIS REST Services}
\usage{
read_edw_lyr(map_name, layer = 0, service = "arcx", target_crs = "EPSG:4326")
}
\arguments{
\item{map_name}{Character. Name of map layer.}

\item{layer}{Integer. Number of layer to read. Default is  zero (0).}

\item{service}{Character. The public ("arcx") or internal ("arcn") ArcGIS
REST Service code. Default is "arcx". You must be on a Forest Service
network for "arcn" to work.}

\item{target_crs}{Coordinate reference system (crs). Default is EPSG:4326
(WGS 84).}
}
\value{
An \link[sf:sf]{sf::sf} object or \link[terra:SpatRaster-class]{terra::SpatRaster}.
}
\description{
This function is not maintained, but remains here for now. Use the
\code{psoGIStools} package. This function reads spatial features from the
Forest Service ArcGIS REST Services. Either the public ArcGIS REST
Service, https://apps.fs.usda.gov/arcx/rest/services/EDW, or the internal
ArcGIS REST Service, https://apps.fs.usda.gov/arcn/rest/services/EDW,
using the \code{arcgislayers} package. You must be on a Forest Service network
to access data from the internal ArcGIS REST Service.
}
\examples{
library(psoSppEvals)

# Administrative Boundary for the Dixie National Forest
admin_bndry <- read_edw_lyr("EDW_ForestSystemBoundaries_01", layer = 1) |> 
  dplyr::filter(forestname == "Dixie National Forest")
}
\seealso{
\code{\link[arcgislayers:arc_read]{arcgislayers::arc_read()}}, \code{\link[sf:st_transform]{sf::st_transform()}}
}
