% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rfss}
\alias{rfss}
\title{2025 Regional Forester's Sensitive Species List for Forest Service Regions 1-10}
\format{
A data frame of 4142 observations and 11 variables.
\describe{
\item{scientific_name}{Scientific name of species.}
\item{common_name}{Common name of species.}
\item{region}{Forest Service Region}
\item{orig_scientific_name}{Origial, uncleaned and unverified scientific name of species.}
\item{taxon_id}{PSO taxon ID.}
\item{gbif_taxonID}{Taxon ID from GBIF.}
\item{kingdom}{Taxonomic Kingdom}
\item{phylum}{Taxonomic Phylum}
\item{class}{Taxonomic Class}
\item{order}{Taxonomic Order}
\item{family}{Taxonomic Family}
\item{genus}{Taxonomic Genus}
\item{species}{Taxonomic Species}
\item{subspecies}{Taxonomic Subspecies}
\item{form}{Taxonomic Form}
\item{variety}{Taxonomic Variety}
}
}
\source{
Compiled from 2025 Regional Forester's Sensitive Species Lists acquired 12 January 2026 from Nichole Panico.
}
\usage{
rfss
}
\description{
2025 Regional Forester's Sensitive Species List for Forest Service Regions 1-10
}
\keyword{datasets}
