% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_pipeline.R
\name{set_up_pipeline}
\alias{set_up_pipeline}
\title{Set Up Species Evaluation Pipeline}
\usage{
set_up_pipeline(dir_path, unit_code)
}
\arguments{
\item{dir_path}{Path to directory you'd like your pipeline.}

\item{unit_code}{Unit acronym.}
}
\value{
Nothing.
}
\description{
This function establishes a targets pipeline for species evaluation data
pulls and evaluation templates. This function copies sub-directories and
files into a specified directroy.
}
\examples{
\dontrun{
library("psoSppEvals")
scc_library <- file.path(Sys.getenv("USERPROFILE"), 
                         "USDA/Mountain Planning Service Group - SCC Library", 
                         "12_Lolo NF")
set_up_pipeline(file.path(scc_library, "Species List"), "LNF")
}
}
