% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_report_functions.R
\name{setup_directories}
\alias{setup_directories}
\title{Create directories for automated evaluations}
\usage{
setup_directories(quarto_params)
}
\arguments{
\item{quarto_params}{Data frame of Quarto parameters from
\code{build_quarto_params()}.}
}
\value{
Nothing.
}
\description{
This function creates the sub-directories for the automated evaluation
reports using the sub-directory paths from \code{build_quarto_params()}.
}
\examples{
\dontrun{
library('psoSppEvals')
spp_list <- get_taxonomies(psoSppEvals::sp_list_ex, correct = TRUE)
qmd_params <- build_quarto_params(spp_list, file.path('output', 'spp_evlas'), 
                                  "Smokey Bear National Forest")
setup_directories(qmd_params)
}
}
\seealso{
\code{\link[=build_quarto_params]{build_quarto_params()}}
}
