% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_functions.R
\name{wkt_string}
\alias{wkt_string}
\title{Internal Function: Create a well-known text string (WTK) string}
\usage{
wkt_string(my_polygon)
}
\arguments{
\item{my_polygon}{An \code{sf} polygon object.}
}
\value{
An \code{sf} vector object.
}
\description{
Creates a well-known text string from a polygon (\code{sf} object). This function
transforms the input polygon to WGS84 prior to calculating the wkt
string.
}
