% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_report_functions.R
\name{write_evals}
\alias{write_evals}
\title{Produce Automated Evaluation Template Documents}
\usage{
write_evals(quarto_params)
}
\arguments{
\item{quarto_params}{Data frame from \code{\link[=build_quarto_params]{build_quarto_params()}}.}
}
\value{
Nothing.
}
\description{
This function runs the Quarto script for each species in the data frame
produced from \code{build_quarto_params()}.
}
\details{
For each species in the data frame, this function runs
\code{quarto::quarto_render()}. This function produces a document in the
current working directory. This function then uses \code{file.copy()} to copy
the document to the output file path in the \emph{output_file} variable. The
document in the current working directory is then deleted using
\code{file.remove()}. On occation, the file in the current working directory
is not removed and will need to be manually deleted.
}
\examples{
\dontrun{
library('mpsgSE')
spp_list <- get_taxonomies(mpsgSE::sp_list_ex, 'scientific_name', 
                           correct = TRUE)
qmd_params <- build_quarto_params(spp_list, file.path('output', 'spp_evlas'), 
                                  "Smokey Bear National Forest")
setup_directories(qmd_params)
write_evals(qmd_params)
}
}
\seealso{
\code{\link[=build_quarto_params]{build_quarto_params()}}, \code{\link[quarto:quarto_render]{quarto::quarto_render()}}, \code{\link[=file.copy]{file.copy()}}, \code{\link[=file.remove]{file.remove()}}
}
