% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_habitat_xlsx.R
\name{write_habitat_xlsx}
\alias{write_habitat_xlsx}
\title{Write Habitat Excel Workbook}
\usage{
write_habitat_xlsx(
  ns_hab_count_data,
  lf_data,
  lf_extract_data,
  unit_code,
  tar_scc_library
)
}
\arguments{
\item{ns_hab_count_data}{Data returned from \code{count_spp_by_hab()}.}

\item{lf_data}{Data returned from \code{pull_landfire()}}

\item{lf_extract_data}{Data returned from \code{extract_landfire_evt()}.}

\item{unit_code}{4-letter Forest Service unit acronym.}

\item{dir_path}{Directory path to save workbook in.}
}
\value{
Nothing.
}
\description{
This function takes the data from \code{count_spp_by_hab()}, \code{pull_landfire()},
and \code{extract_landfire_evt()} and writes them to an Excel workbook.
}
\examples{
# Coming soon, to a theater near you...
message("Check back soon")
}
\seealso{
\code{\link[=count_spp_by_hab]{count_spp_by_hab()}}, \code{\link[=pull_landfire]{pull_landfire()}}, \code{\link[=extract_landfire_evt]{extract_landfire_evt()}}
}
