% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_range_data.R
\name{write_range_data}
\alias{write_range_data}
\title{Write range data to geodatabase}
\usage{
write_range_data(ebird_range, iucn_maps, bien_maps, gdb_path)
}
\arguments{
\item{ebird_range}{eBird range maps from this pipeline}

\item{iucn_maps}{iucn maps from this pipeline}

\item{bien_maps}{bien maps from this pipeline}

\item{gdb_path}{path to geodatabase}
}
\value{
Nothing.
}
\description{
Write range data to geodatabase
}
\examples{
# Coming soon 
message("Stay tuned.")

}
\seealso{
\code{\link[=download_ebird_range_maps]{download_ebird_range_maps()}}, \code{\link[=download_bien_range_maps]{download_bien_range_maps()}}, \code{\link[=build_iucn_maps]{build_iucn_maps()}}
}
