% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_report_functions.R
\name{build_quarto_params}
\alias{build_quarto_params}
\title{Build Quarto Parameters for Automated Evaluation Templates}
\usage{
build_quarto_params(
  spp_list,
  fs_unit_name,
  states,
  crs,
  output_path = file.path("output/spp_evals")
)
}
\arguments{
\item{spp_list}{Species list with taxon ID's and taxonomies from
\code{get_taxonomies()}.}

\item{fs_unit_name}{Character string for Forest Service unit name.}

\item{states}{Character string or a concatenated vector for states the forest
or grassland overlaps.}

\item{crs}{Coordinate reference system used to mapping.}

\item{output_path}{The path to the directory where you want the reports saved.
This can be}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function builds the Quarto parameters to produce the automated species
evaluations.
}
\details{
This function makes a data frame with the following variables:
\itemize{
\item taxon_id: taxonomic ID from 'get_taxonomies()'
\item scientific_name: scientific name
\item common_name: common name
\item file_name: name of Microsoft Word document (.docx)
\item subfolder_path: sub-folder path
\item full_path: file path for output file
\item unit_name: Forest Service unit name
}

The sub-folder paths is assigned using taxonomy.
\itemize{
\item Taxa in the taxonomic Kingdom Plantae are placed in the Plants folder.
\item Taxa in the taxonomic Phylum Ascomycota are placed in the Lichens folder.
\item Taxa in the taxonomic Phylum Arthropoda and Mollusca are placed in the Invertebrates folder.
\item Taxa in the taxonomic Class Aves are placed in the Lichens folder.
\item Taxa in the taxonomic Classes Amphibia, Squamata, and Testudines are placed in the Herpetofauna folder.
\item Taxa in the taxonomic Class Mammalia are placed in the Mammals folder.
\item Taxa in the do not have a taxonomic class are placed in the Fishes folder. For some reason Fishes are returned with Class == NA with \code{get_taxonomies()}.
}
}
\examples{
\dontrun{
library('psoSppEvals')
spp_list <- get_taxonomies(psoSppEvals::sp_list_ex, correct = TRUE)
qmd_params <- build_quarto_params(spp_list, file.path('output', 'spp_evlas'))
}
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}
}
