% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{name_corrections}
\alias{name_corrections}
\title{Taxon ID and Scientific Name corrections}
\format{
A data frame of 7 observations and 4 variables.
\describe{
\item{common_name}{Common species name.}
\item{errored_name}{Name that is not being queried in GBIF.}
\item{corrected_name}{A scientific name for the same speces in the GBIF backbome taxononmy.}
\item{taxon_id}{Taxon ID.}
\item{gbif_taxonID}{Taxon ID from GBIF.}
\item{kingdom}{Taxonomic Kingdom}
\item{phylum}{Taxonomic Phylum}
\item{class}{Taxonomic Class}
\item{order}{Taxonomic Order}
\item{family}{Taxonomic Family}
\item{genus}{Taxonomic Genus}
\item{species}{Taxonomic Species}
\item{subspecies}{Taxonomic Subspecies}
}
}
\usage{
name_corrections
}
\description{
Manually generated through trial and error
}
\keyword{datasets}
