% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_range_data.R
\name{write_range_data}
\alias{write_range_data}
\title{Write range data to geodatabase}
\usage{
write_range_data(
  bien_maps,
  ebird_range,
  iucn_maps,
  gdb_path,
  return_sf = FALSE
)
}
\arguments{
\item{bien_maps}{bien maps from this pipeline}

\item{ebird_range}{eBird range maps from this pipeline}

\item{iucn_maps}{iucn maps from this pipeline}

\item{gdb_path}{path to geodatabase}

\item{return_sf}{Optional. TRUE/FALSE. Return a list of \code{sf} objects. Default
is FALSE.}
}
\value{
A list of \code{sf} objects including BIEN, eBird, and IUCN range data.
}
\description{
This function writes spatial (\code{sf}) BIEN, eBird, and IUCN range data to a
geodatabase and has the option to return a list of \code{sf} objects.
}
\examples{
# Coming soon 
message("Stay tuned.")

}
\seealso{
\code{\link[=download_bien_maps]{download_bien_maps()}}, \code{\link[=download_ebird_range_maps]{download_ebird_range_maps()}}, \code{\link[=build_iucn_maps]{build_iucn_maps()}}
}
